RUE-1831: split canonical comptime engine modules
