+++ crates/rue-air/src/lib.rs
@@ -87,16 +87,18 @@ pub use sema::{
     BodyFactProvider, BodyNamedDependencyEvent, BodyOwnerEndpoint, BodyOwnerKind, BodyOwnerToken,
     BodyRirBundle, BodyRirIndexAttribution, BodyRirView, BuiltinTypeCallHead,
     ComptimeAnonymousKind, ComptimeArgMode, ComptimeArrayLengthBinding, ComptimeCallAdmission,
-    ComptimeCallArgument, ComptimeCallKey, ComptimeCallMemoLookup, ComptimeCallPreparation,
-    ComptimeCompletedCallMemo, ComptimeEngine, ComptimeEnv, ComptimeExpressionIntrinsic,
-    ComptimeExpressionIntrinsicRequest, ComptimeField, ComptimeFile, ComptimeFrame, ComptimeHost,
-    ComptimeHostError, ComptimeHostResult, ComptimeIdentity, ComptimeIntegerBound,
-    ComptimeMemoInsertError, ComptimeMemoizedOutcome, ComptimeMethodDescriptor,
-    ComptimeMethodParameter, ComptimeMethodType, ComptimeName, ComptimeNamedValueResolution,
-    ComptimeOutcome, ComptimeProgram, ComptimeProgramKey, ComptimeProgramRegistrationError,
-    ComptimeProgramRegistry, ComptimeSite, ComptimeSiteKind, ComptimeStructuredTypeResolution,
-    ComptimeStructuredTypeSuspension, ComptimeTargetIntrinsic, ComptimeTrap, ComptimeType,
-    ComptimeTypeIntrinsic, ComptimeValue, ConstInfo, ConstValue, DeclarationBindingWork,
+    ComptimeCallArgument, ComptimeCallHost, ComptimeCallKey, ComptimeCallMemoLookup,
+    ComptimeCallPreparation, ComptimeCompletedCallMemo, ComptimeEngine, ComptimeEnv,
+    ComptimeExpressionIntrinsic, ComptimeExpressionIntrinsicRequest, ComptimeField, ComptimeFile,
+    ComptimeFrame, ComptimeHost, ComptimeHostError, ComptimeHostResult, ComptimeHostTypes,
+    ComptimeIdentity, ComptimeIntegerBound, ComptimeIntrinsicHost, ComptimeMemoInsertError,
+    ComptimeMemoizedOutcome, ComptimeMethodDescriptor, ComptimeMethodParameter, ComptimeMethodType,
+    ComptimeName, ComptimeNamedValueResolution, ComptimeOutcome, ComptimeProgram,
+    ComptimeProgramHost, ComptimeProgramKey, ComptimeProgramRegistrationError,
+    ComptimeProgramRegistry, ComptimeSite, ComptimeSiteKind, ComptimeStructuredTypeHost,
+    ComptimeStructuredTypeResolution, ComptimeStructuredTypeSuspension, ComptimeTargetIntrinsic,
+    ComptimeTrap, ComptimeType, ComptimeTypeHost, ComptimeTypeIntrinsic, ComptimeValue,
+    ComptimeValueHost, ConstInfo, ConstValue, DeclarationBindingWork,
     DeclarationBuiltinTypeCallHeadDependencyEvent, DeclarationTypeCallHeadDependencyEvent,
     DeclarationTypeDependencyEvent, DeclarationTypeDependencyKind,
     DeclarationTypeDependencySourceKind, DeclarationTypeDependencyTargetKind, DropCopyMetadata,
